requirements.txt
faiss-cpu
librosa
nltk
numpy
torch>=1.9.0
transformers>=4.31.0
torchaudio>=0.10.0